requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# WhatsApp Cloud API
pywa==3.9.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
aiosqlite==1.3.0

# Cache / queue
redis==5.2.1
celery==5.4.0

# Config
pydantic-settings==2.7.0
python-dotenv==1.0.1

# HTTP client
httpx==0.28.1

# Security & JWT
PyJWT==2.12.1

# Ingestion (mandi price pipeline)
beautifulsoup4==4.12.3
lxml==5.3.0
tenacity==9.0.0

# LLM fallback
google-generativeai==0.8.3

# Image handling (lightweight; needed for media message parsing)
pillow>=10.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
